Attempting to Webpack Bootstrap Issues
At the moment the Bootstrap styles are loading alright but functionality like Collapse etc are throwing JS errors. This is only happening when using Webpack pack files. If the code makes direct references to the Bootstrap JS and jQuery files on the CDNs then everything works file. This leads me to believe that Webpack is probably not configured properly for the app. This is resulting in Bootstrap functionality with jQuery not working as expected.