Webpack Bootstrap Issue Fixed
Fixed the issue by using an older version of jQuery (3.4.1) instead of the newer one. Experimented with using jQuery-Slim instead and that caused a whole lot of Webpack compiling issues. May have to revisit that at a later time. At the moment using all of Bootstraps modules. Will try to refine it at a later point as we don't need all of Bootstrap modules for the website.